move wrapper to App component
