+++ src/pages/App.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Route, Switch } from 'react-router-dom';
+import { Route, Switch, withRouter } from 'react-router-dom';
 
 import Header from 'legacy/work/Header';
 import Footer from 'legacy/work/Footer';
@@ -29,34 +29,50 @@ import '../styled/customStyles.scss';
 import HostListingEdit from 'legacy/routes/Host/HostListing/HostListingEdit';
 import HostListingCalendar from 'legacy/routes/Host/HostListingCalendar';
 
+class ScrollToTop extends React.Component<RouterProps, {}> {
+  componentDidUpdate(prevProps: RouterProps) {
+    if (this.props.location.pathname !== prevProps.location.pathname) {
+      window.scrollTo(0, 0);
+    }
+  }
+
+  render() {
+    return this.props.children;
+  }
+}
+
+const ScrollToTopWithRouter = withRouter(ScrollToTop);
+
 const Work = () => (
   <div className="min-height-100vh">
     <Header />
     <div className="min-height-100vh">
-      <Switch>
-        <Route path="/about" component={About} />
-        <Route exact path="/account/stripe/link" component={StripeLink} />
-        <Route exact path="/account/stripe/new" component={StripeNew} />
-        <Route exact path="/account/stripe/complete" component={StripeComplete} />
-        <Route path="/account/action" component={FirebaseEmailHandler} />
-        <Route path="/account" component={Account} />
-        <Route path="/markets" component={Markets} />
-        <AuthenticatedRoute path="/bookings" component={Booking} />
-        <Route path="/forgot_password" component={ForgotPassword} />
-        <AuthenticatedRoute path="/host/listings/:id/calendar" component={HostListingCalendar} />
-        <AuthenticatedRoute path="/host/listings/:id" render={(props: RouterProps) => <HostListingEdit {...props} />} />
-        <AuthenticatedRoute path="/host" component={Host} />
-        <Route exact path="/listings/:id" component={Listing} />
-        <Route exact path="/login" component={Login} />
-        <Route exact path="/logout" component={Logout} />
-        <Route exact path="/search" component={Search} />
-        <Route exact path="/signup" component={Signup} />
-        <AuthenticatedRoute exact path="/trips/:id/receipt" component={TripsReceipt} />
-        <AuthenticatedRoute path="/trips" component={Trips} />
-        <Route path="/hosts/signup" component={HostSignup} />
-        <Route exact path="/" component={Home} />
-        <Route path="/" component={NotFound} />
-      </Switch>
+      <ScrollToTopWithRouter>
+        <Switch>
+          <Route path="/about" component={About} />
+          <Route exact path="/account/stripe/link" component={StripeLink} />
+          <Route exact path="/account/stripe/new" component={StripeNew} />
+          <Route exact path="/account/stripe/complete" component={StripeComplete} />
+          <Route path="/account/action" component={FirebaseEmailHandler} />
+          <Route path="/account" component={Account} />
+          <Route path="/markets" component={Markets} />
+          <AuthenticatedRoute path="/bookings" component={Booking} />
+          <Route path="/forgot_password" component={ForgotPassword} />
+          <AuthenticatedRoute path="/host/listings/:id/calendar" component={HostListingCalendar} />
+          <AuthenticatedRoute path="/host/listings/:id" render={(props: RouterProps) => <HostListingEdit {...props} />} />
+          <AuthenticatedRoute path="/host" component={Host} />
+          <Route exact path="/listings/:id" component={Listing} />
+          <Route exact path="/login" component={Login} />
+          <Route exact path="/logout" component={Logout} />
+          <Route exact path="/search" component={Search} />
+          <Route exact path="/signup" component={Signup} />
+          <AuthenticatedRoute exact path="/trips/:id/receipt" component={TripsReceipt} />
+          <AuthenticatedRoute path="/trips" component={Trips} />
+          <Route path="/hosts/signup" component={HostSignup} />
+          <Route exact path="/" component={Home} />
+          <Route path="/" component={NotFound} />
+        </Switch>
+      </ScrollToTopWithRouter>
     </div>
     <Footer />
   </div>
